smartGNU/requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
ecdsa==0.15
esptool==2.8
firebase-admin==3.2.1
gcloud==0.17.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
httplib2==0.17.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.3
jws==0.1.3
MarkupSafe==1.1.1
msgpack==0.6.2
oauth2client==3.0.0
openapi-codec==1.3.2
paho-mqtt==1.5.0
Pillow==7.0.0
protobuf==3.11.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.4.3
PyHamcrest==1.10.1
pyOpenSSL==19.1.0
Pyrebase==3.0.27
pyserial==3.4
python-jwt==2.0.1
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
service-identity==18.1.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
uritemplate==3.0.1
urllib3==1.25.8
zope.interface==4.7.1
